Move Encoder, Decoder, DatumReader and DatumWriter interfaces to cdap-formats
