+++ cdap-common/src/main/java/co/cask/cdap/internal/io/ASMDatumWriterFactory.java
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2014 Cask Data, Inc.
+ * Copyright © 2014-2017 Cask Data, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
@@ -17,6 +17,7 @@
 package co.cask.cdap.internal.io;
 
 import co.cask.cdap.api.data.schema.Schema;
+import co.cask.cdap.common.io.DatumWriter;
 import co.cask.cdap.internal.asm.ByteCodeClassLoader;
 import co.cask.cdap.internal.asm.ClassDefinition;
 import com.google.common.base.Objects;
