+++ cdap-formats/src/main/java/co/cask/cdap/common/io/DatumReader.java
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2014 Cask Data, Inc.
+ * Copyright © 2017 Cask Data, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
@@ -13,10 +13,9 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package co.cask.cdap.internal.io;
+package co.cask.cdap.common.io;
 
 import co.cask.cdap.api.data.schema.Schema;
-import co.cask.cdap.common.io.Decoder;
 
 import java.io.IOException;
