+++ cdap-formats/src/main/java/co/cask/cdap/common/io/DatumWriter.java
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2014 Cask Data, Inc.
+ * Copyright © 2017 Cask Data, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
@@ -14,9 +14,7 @@
  * the License.
  */
 
-package co.cask.cdap.internal.io;
-
-import co.cask.cdap.common.io.Encoder;
+package co.cask.cdap.common.io;
 
 import java.io.IOException;
