+++ cdap-security/src/main/java/co/cask/cdap/security/auth/AccessTokenIdentifierCodec.java
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2014 Cask Data, Inc.
+ * Copyright © 2014-2017 Cask Data, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
@@ -20,11 +20,11 @@ import co.cask.cdap.api.data.schema.Schema;
 import co.cask.cdap.common.io.BinaryDecoder;
 import co.cask.cdap.common.io.BinaryEncoder;
 import co.cask.cdap.common.io.Codec;
+import co.cask.cdap.common.io.DatumReader;
+import co.cask.cdap.common.io.DatumWriter;
 import co.cask.cdap.common.io.Decoder;
 import co.cask.cdap.common.io.Encoder;
-import co.cask.cdap.internal.io.DatumReader;
 import co.cask.cdap.internal.io.DatumReaderFactory;
-import co.cask.cdap.internal.io.DatumWriter;
 import co.cask.cdap.internal.io.DatumWriterFactory;
 import com.google.common.reflect.TypeToken;
 import com.google.inject.Inject;
