+++ cdap-watchdog/src/main/java/co/cask/cdap/metrics/process/MetricsMessageCallback.java
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2014-2016 Cask Data, Inc.
+ * Copyright © 2014-2017 Cask Data, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
@@ -23,7 +23,7 @@ import co.cask.cdap.api.metrics.MetricValue;
 import co.cask.cdap.api.metrics.MetricValues;
 import co.cask.cdap.api.metrics.MetricsContext;
 import co.cask.cdap.common.io.BinaryDecoder;
-import co.cask.cdap.internal.io.DatumReader;
+import co.cask.cdap.common.io.DatumReader;
 import co.cask.common.io.ByteBufferInputStream;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.Lists;
